Stop rendering merged courses in expression--course
We have to merge the course query and found course objects when we
encounter them, but that led to rendering all of the information from
the merged course in the block.

Which felt to me like it implied that you had to have taken this
specific course, when all it meant was that this course happened to
match.

Anyway. That's solved now.