add removeDuplicateAreas test
This _also_ flips the logic for which area to remove when there is a duplicate.

Before, we would remove all but the shortest-pathed area, as sorted by sourcePath.length.
Having studied it more, that seems backwards: we should be removing all but the _longest_
sourcePath.

That solves a situation such as when a major adds a new revision, when it hadn't
had any revisions before, so the stored major would then exist in two places:
both path.yaml, and path-revision.yaml.